Setup the addon system for true implementation - Requires store to complete Also modularized the addon system TG-59 #done TG-24 #done
