+++ src/app.container.ts
@@ -1,11 +1,11 @@
 import 'reflect-metadata'
 import { container } from 'inversify-props'
 import IHomeDashboardService from '@/components/home-dashboard/services/IHomeDashboardService'
-import IAddonsService from '@/services/interfaces/IAddonsService'
+import IAddonsService from '@/components/add-on-manager/services/IAddonsService'
 import ICaseFileService from '@/components/case-file/services/ICaseFileService'
 import INetworkMonitorService from '@/components/network-monitor/services/INetworkMonitorService'
 import ICameraMonitorService from '@/components/camera-monitor/services/ICameraMonitorService'
-import AddonsService from '@/services/implementations/AddonsService'
+import AddonsService from '@/components/add-on-manager/services/AddonsService'
 import CaseFileService from '@/components/case-file/services/CaseFileService'
 import HomeDashboardService from '@/components/home-dashboard/services/HomeDashboardService'
 import NetworkMonitorService from '@/components/network-monitor/services/NetworkMonitorService'
