--- src/components/add-on-manager/AddOnsCard.vue
@@ -1,17 +1,0 @@
-<template>
-  <v-card>
-    <p>Add Ons</p>
-  </v-card>
-</template>
-
-<script lang="ts">
-import Vue from 'vue'
-import Component from 'vue-class-component'
-
-@Component({
-  name: 'AddOnsCard'
-})
-export default class AddOnsCard extends Vue {
-
-}
-</script>
