+++ src/components/add-on-manager/services/IAddonsService.ts
@@ -1,0 +1,7 @@
+import IService from '../../../services/interfaces/IService'
+import IMenuItem from '@/types/IMenuItem'
+
+export default interface IAddonsService extends IService {
+    getEnabledAddons(): IMenuItem[]
+    enableAddon(addOnModel: IMenuItem): void
+}
