--- src/services/implementations/AddonsService.ts
@@ -1,19 +1,0 @@
-import 'reflect-metadata'
-import { Vue, Component, Model } from 'vue-property-decorator'
-import MenuItemService from '@/services/implementations/MenuItemService'
-import IMenuItem from '@/types/IMenuItem'
-import IService from '@/services/interfaces/IService'
-import { injectable } from 'inversify-props'
-
-@injectable()
-export default class AddonsService extends MenuItemService implements IService {
-  defaultModel () : IMenuItem {
-    const properties : IMenuItem = {
-      title: 'Add Ons',
-      icon: 'mdi-view-grid-plus',
-      path: '/add-ons'
-    }
-    this.setMenuItemProperties(properties)
-    return this.getModel()
-  }
-}
