--- src/services/interfaces/IAddonsService.ts
@@ -1,4 +1,0 @@
-import IService from './IService'
-
-export default interface IAddonsService extends IService {
-}
